# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  levelarea.CreateRequest:
    properties:
      below_level_area_id:
        minimum: 1
        type: integer
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 20
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - name
    type: object
  levelarea.UpdateRequest:
    properties:
      below_level_area_id:
        minimum: 1
        type: integer
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 20
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - name
    type: object
  levelvolunteer.CreateRequest:
    properties:
      below_level_volunteer_id:
        minimum: 1
        type: integer
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 20
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - name
    type: object
  levelvolunteer.UpdateRequest:
    properties:
      below_level_volunteer_id:
        minimum: 1
        type: integer
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 20
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - name
    type: object
  masteractivity.CreateRequest:
    properties:
      category_activity_id:
        minimum: 1
        type: integer
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      target:
        minimum: 0
        type: integer
    required:
    - category_activity_id
    - name
    - target
    type: object
  masteractivity.UpdateRequest:
    properties:
      category_activity_id:
        minimum: 1
        type: integer
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      target:
        minimum: 0
        type: integer
    required:
    - category_activity_id
    - name
    - target
    type: object
...
info:
  contact: {}
  description: API documentation
  title: Social BE API
  version: "1.0"
paths:
...
  /master/level-areas:
    get:
      description: Get list of level area
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all level area
      tags:
      - level-area
    post:
      consumes:
      - application/json
      description: Create new level area
      parameters:
      - description: Create level area
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/levelarea.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create level area
      tags:
      - level-area
  /master/level-areas/{id}:
    delete:
      description: Soft delete level area
      parameters:
      - description: Level Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete level area
      tags:
      - level-area
    get:
      description: Get detail level area
      parameters:
      - description: Level Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get level area by ID
      tags:
      - level-area
    put:
      consumes:
      - application/json
      description: Update existing level area
      parameters:
      - description: Level Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update level area
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/levelarea.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update level area
      tags:
      - level-area
  /master/level-volunteers:
    get:
      description: Get paginated list of level volunteer
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all level volunteer
      tags:
      - level-volunteer
    post:
      consumes:
      - application/json
      description: Create new level volunteer
      parameters:
      - description: Create level volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/levelvolunteer.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create level volunteer
      tags:
      - level-volunteer
  /master/level-volunteers/{id}:
    delete:
      description: Soft delete level volunteer
      parameters:
      - description: Level Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete level volunteer
      tags:
      - level-volunteer
    get:
      description: Get detail level volunteer
      parameters:
      - description: Level Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get level volunteer by ID
      tags:
      - level-volunteer
    put:
      consumes:
      - application/json
      description: Update existing level volunteer
      parameters:
      - description: Level Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update level volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/levelvolunteer.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update level volunteer
      tags:
      - level-volunteer
  /master/master-activities:
    get:
      description: Get paginated list of master activity
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all master activity
      tags:
      - master-activity
    post:
      consumes:
      - application/json
      description: Create new master activity
      parameters:
      - description: Create master activity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masteractivity.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create master activity
      tags:
      - master-activity
  /master/master-activities/{id}:
    delete:
      description: Soft delete master activity
      parameters:
      - description: Master Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete master activity
      tags:
      - master-activity
    get:
      description: Get detail master activity
      parameters:
      - description: Master Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get master activity by ID
      tags:
      - master-activity
    put:
      consumes:
      - application/json
      description: Update existing master activity
      parameters:
      - description: Master Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update master activity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masteractivity.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update master activity
      tags:
      - master-activity
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"